Update Storage protocol to return state and value in get
Although it seems like a bit of an odd choice that get returns the store and the value at the reference, I think there's a lot of benefits to this approach. For a start, GenServer requires a new state when handling the :get call anyways, so we can unpack the store and value before repacking it into the handle_call return value. That means we can have stores that mutate on a get, like a CachingStore changing the fast store, or we can have stores that don't change on a get (like a regular MapStore).